requirements.txt
scrapy>=2.11.0
pandas>=2.1.0
scikit-learn>=1.3.0
pymongo>=4.5.0
numpy>=1.24.0
matplotlib>=3.7.0
pytest>=7.4.0
flake8>=6.1.0
mypy>=1.5.0
python-dotenv>=1.0.0
Flask>=2.3.0
SQLAlchemy>=2.0.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
pre-commit>=3.4.0
jupyter>=1.0.0
python-dateutil>=2.8.2
ta-lib>=0.4.28
